% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermittent_ata.r
\name{components.intermittentATA}
\alias{components.intermittentATA}
\title{Extract estimated states from an intermittentATA model.}
\usage{
\method{components}{intermittentATA}(object, ...)
}
\arguments{
\item{object}{An estimated model.}

\item{...}{Unused.}
}
\value{
A [fabletools::dable()] containing estimated states.
}
\description{
Extract estimated states from an intermittentATA model.
}
\examples{
library(intermittentATA)
as_tsibble(fmcgData) \%>\% 
   model(crostonata = intermittentATA(value ~ d_trend(type = "M", parQ = 1) + i_trend("A") + intermittent("croston"))) \%>\% components()

}
