% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermittent_ata.r
\name{fitted.intermittentATA}
\alias{fitted.intermittentATA}
\title{Extract fitted values}
\usage{
\method{fitted}{intermittentATA}(object, ...)
}
\arguments{
\item{object}{An estimated model.}

\item{...}{Unused.}
}
\value{
A vector of fitted values.
}
\description{
Extracts the fitted values from an estimated intermittentATA model.
}
