% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermittent_ata.r
\name{format.intermittentATA}
\alias{format.intermittentATA}
\title{Format of intermittentATA model}
\usage{
\method{format}{intermittentATA}(x, ...)
}
\arguments{
\item{x}{An estimated model.}

\item{...}{Unused.}
}
\value{
The forecasting model's name.
}
\description{
Format of intermittentATA model
}
