% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermittent_ata.r
\name{model_sum.intermittentATA}
\alias{model_sum.intermittentATA}
\title{Summary of intermittentATA model}
\usage{
\method{model_sum}{intermittentATA}(x, ...)
}
\arguments{
\item{x}{An estimated model.}

\item{...}{Unused.}
}
\value{
The model's summary specs.
}
\description{
Summary of intermittentATA model
}
