% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermittent_ata.r
\name{report.intermittentATA}
\alias{report.intermittentATA}
\title{Specialized Screen Print Function of intermittentATA model}
\usage{
\method{report}{intermittentATA}(object, ...)
}
\arguments{
\item{object}{An estimated model.}

\item{...}{Unused.}
}
\value{
a summary for the results of the ATAforecasting
}
\description{
Specialized Screen Print Function of intermittentATA model
}
\examples{
 library(intermittentATA)
 as_tsibble(fmcgData) \%>\%
     model(crostonata = intermittentATA(value ~ d_trend(type = "M", parQ = 1) + i_trend("A") + intermittent("croston"))) \%>\% report()

}
