% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermittent_ata.r
\name{residuals.intermittentATA}
\alias{residuals.intermittentATA}
\title{Extract model residuals}
\usage{
\method{residuals}{intermittentATA}(object, ...)
}
\arguments{
\item{object}{An estimated model.}

\item{...}{Unused.}
}
\value{
A vector of residuals.
}
\description{
Extracts the residuals from an estimated intermittentATA model.
}
