% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intermittent_ata.r
\name{tidy.intermittentATA}
\alias{tidy.intermittentATA}
\title{Tidy a intermittentATA model}
\usage{
\method{tidy}{intermittentATA}(x, ...)
}
\arguments{
\item{x}{An estimated model.}

\item{...}{Unused.}
}
\value{
The model's coefficients in a `tibble`.
}
\description{
Tidy a intermittentATA model
}
\examples{
library(intermittentATA)
as_tsibble(fmcgData) \%>\% 
    model(crostonata = intermittentATA(value ~ d_trend(type = "M", parQ = 1) + i_trend("A") + intermittent("croston"))) \%>\% tidy()

}
